Make labels optional
