+++ charts/core/src/Chart.stories.tsx
@@ -55,8 +55,8 @@ export default {
     eventMarkerLineLevel: 'warning',
     eventMarkerLinePosition: new Date('2024-01-01T00:20:00'),
     renderEventMarkerPoint: true,
-    eventMarkerPointMessage: 'Event marker line message',
-    eventMarkerPointLabel: 'Event marker line label',
+    eventMarkerPointMessage: 'Event marker point message',
+    eventMarkerPointLabel: 'Event marker point label',
     eventMarkerPointLevel: 'warning',
     eventMarkerPointXPosition: new Date('2024-01-01T00:37:00'),
     eventMarkerPointYPosition: 699,
