+++ charts/core/src/ThresholdLine/ThresholdLine.types.ts
@@ -1,5 +1,12 @@
+import { Theme } from '@leafygreen-ui/lib';
+
 export interface ThresholdLineProps {
   position: number;
-  label: string;
   value: string;
+  label?: string;
+}
+
+export interface GetThresholdLineConfig extends ThresholdLineProps {
+  name: string;
+  theme: Theme;
 }
